Configures API for development with Scalar
Configures the API to use the Scalar OpenAPI tool for development.
This includes adding the Scalar.AspNetCore package, mapping the Scalar API reference endpoint, and updating the API base URL in the client application and launch settings to port 5001 for HTTPS and 5000 for HTTP.
The changes improve the developer experience by providing a user-friendly interface for exploring and testing the API.
Also removes a no longer required class and http configuration file.